+++ packages/react-html/CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
-- Added metadata components: `Baseline`, `AppleHomeScreen` and `Responsive`. [#481](https://github.com/Shopify/quilt/pull/481)
+- Added metadata components: `AppleHomeScreen` and `Responsive`. [#481](https://github.com/Shopify/quilt/pull/481)
 
 ## 6.0.2 - 2019-01-09
