+++ packages/react-html/src/components/index.ts
@@ -7,4 +7,3 @@ export {default as Style} from './Style';
 export {default as Title} from './Title';
 export {default as Responsive} from './Responsive';
 export {default as AppleHomeScreen} from './AppleHomeScreen';
-export {default as Baseline} from './Baseline';
