--- packages/react-html/src/tests/utilities.ts
@@ -1,9 +1,0 @@
-export function findMeta(wrapper, props) {
-  return wrapper.findWhere(component => {
-    const propsArray = Object.keys(props);
-    return (
-      propsArray.filter(prop => component.prop(prop) === props[prop]).length ===
-      propsArray.length
-    );
-  });
-}
